refactor(matrix-utils-auth): split tests into domain-specific modules
reorganize auth tests into domain-specific modules and update the testing
guide to reflect the new test layout and files
